+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Support for reading and validating digital signatures
 * Output info regarding digital signatures when using the `hexapdf info` command
+* [HexaPDF::Type::AcroForm::Form#create_signature_field] for adding signature
+  fields
 
 ### Changed